fix(awaitHeight): allow to wait for arbitrary number of blocks
`awaitHeight` was accepting the amount of attempts by default equal to 20 that actually should
depend on a difference between current and desired heights. Usually we assume that used node is
being synced therefore any height should be reachable. We are not doing a real-time operating
system, so we don't care about riching a height in a limited time.