feat(chain)!: add `addAccounts` option instead address in `txDryRun`
BREAKING CHANGE: `txDryRun` accepts account balances in options
Apply a change
```diff
-txDryRun(tx, address)
+txDryRun(tx, { addAccounts: [{ address, amount: n }] })
```
Where `amount` is a value in aettos to add to the on-chain balance of that account.
Alternatively, `addAccounts` can be omitted to use the on-chain balance
```js
txDryRun(tx)
```
Where `amount` is a value in aettos to add to the on-chain balance of that account.